implementing skeleton update

+++ src/engine/time.rs
@@ -1,5 +1,5 @@
 #[derive(Default, Clone)]
 pub struct Time {
     pub dt: f32,
-    pub time: f32,
+    pub time: f64,
 }
